Remove `test_support` feature flags
