+++ tokio-trace-core/Cargo.toml
@@ -3,12 +3,5 @@ name = "tokio-trace-core"
 version = "0.1.0"
 authors = ["Eliza Weisman <eliza@buoyant.io>"]
 
-[features]
-default = []
-test-support = []
-
 [dependencies]
 lazy_static = "1.0.0"
-
-[dev-dependencies]
-futures = "0.1"
