+++ tokio-trace-futures/Cargo.toml
@@ -8,5 +8,4 @@ futures = "0.1"
 tokio-trace = { path = "../tokio-trace" }
 
 [dev-dependencies]
-tokio-trace = { path = "../tokio-trace", features = ["test-support"] }
 tokio = "0.1"
