+++ tokio-trace/tests/support/mod.rs
@@ -1,0 +1,3 @@
+#![allow(dead_code)]
+pub mod span;
+pub mod subscriber;
